requirements.txt
setuptools
ipykernel
jupyter
matplotlib
numpy
sklearn
pandas
Keras
flask
requests
grpcio-tools
nltk
hdbscan
click
git+https://github.com/NewKnowledge/tslearn@932273f710fde4eca13707faf42945cdeb82e84f
git+https://github.com/NewKnowledge/rrcf@1844465f28816b55ef4ef481809dcf26f968c5c3
git+https://github.com/NewKnowledge/simon@81abbf365a8bb0b4e73be1e0cbb715d1654bd15b
git+https://github.com/NewKnowledge/sloth@b9b7b121a1075267dec1eedf5dea94be19de0e4c